TourInfo 권한 ADMIN, DEV 만 주어짐 TourInfo user_id가 일치해야지만 글 수정, 삭제하도록 변경

+++ src/main/java/com/landmark/app/service/impl/TourInfoServiceImpl.java
@@ -97,30 +97,31 @@ public class TourInfoServiceImpl extends LoggerUtils implements TourInfoService 
      * 관광지 수정
      */
     @Override
-    public void updateTours(TourInfoDTO tourInfoDTO){
+    public void updateTours(TourInfoDTO tourInfoDTO, int getuserId){
         try{
-            String addr1 = tourInfoDTO.getAddr1();
-            String addr2 = tourInfoDTO.getAddr2();
-            int areaCode = tourInfoDTO.getAreaCode();
-            int sigunguCode= tourInfoDTO.getSigunguCode();
-            String cat1 = tourInfoDTO.getCat1();
-            String cat2 = tourInfoDTO.getCat2();
-            String cat3 = tourInfoDTO.getCat3();
-            int contentId = tourInfoDTO.getContentId();
-            int contentTypeId = tourInfoDTO.getContentTypeId();
-            String tel = tourInfoDTO.getTel();
-            String title = tourInfoDTO.getTitle();
-            String overview = tourInfoDTO.getOverview();
-            Date modifiedTime = new Date();
-            int id = tourInfoDTO.getId();
             int userId = tourInfoDTO.getUserId();
+            if(getuserId == userId){
+                String addr1 = tourInfoDTO.getAddr1();
+                String addr2 = tourInfoDTO.getAddr2();
+                int areaCode = tourInfoDTO.getAreaCode();
+                int sigunguCode= tourInfoDTO.getSigunguCode();
+                String cat1 = tourInfoDTO.getCat1();
+                String cat2 = tourInfoDTO.getCat2();
+                String cat3 = tourInfoDTO.getCat3();
+                int contentId = tourInfoDTO.getContentId();
+                int contentTypeId = tourInfoDTO.getContentTypeId();
+                String tel = tourInfoDTO.getTel();
+                String title = tourInfoDTO.getTitle();
+                String overview = tourInfoDTO.getOverview();
+                Date modifiedTime = new Date();
+                int id = tourInfoDTO.getId();
 
-            tourInfoRepository.updateTourByUserId(addr1, addr2, areaCode, sigunguCode,
-                    cat1, cat2, cat3,
-                    contentId,contentTypeId,
-                    tel, title, overview, modifiedTime,
-                    id, userId);
-
+                tourInfoRepository.updateTourByUserId(addr1, addr2, areaCode, sigunguCode,
+                        cat1, cat2, cat3,
+                        contentId,contentTypeId,
+                        tel, title, overview, modifiedTime,
+                        id, userId);
+            }
         } catch (Exception e){
             logger.error("updateTours : " + e.getMessage());
         }
@@ -128,25 +129,15 @@ public class TourInfoServiceImpl extends LoggerUtils implements TourInfoService 
 
 
     /**
-     * 관광지 조회
-     */
-    @Override
-    public List<TourInfoDTO.RegisteredTourInfoDTO> getTours(int user_id){
-        try{
-
-        } catch (Exception e){
-            return null;
-        }
-        return null;
-    }
-
-    /**
      * 관광지 삭제
      */
     @Override
     public void deleteTours(int id, int user_id){
         try{
-            tourInfoRepository.deleteTourByIdUserId(id, user_id);
+            TourInfoDTO tourInfoDTO = TourInfoDTO.of(tourInfoRepository.findByIdUserId(id, user_id));
+            if(tourInfoDTO.getUserId() == user_id){
+                tourInfoRepository.deleteTourByIdUserId(id, user_id);
+            }
         } catch (Exception e){
             logger.error("deleteByUserIdId" + e.getMessage());
         }
